requirements.txt
# Core Django
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
django-allauth==0.61.1
django-otp==1.4.1
qrcode==7.4.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Phone number handling
phonenumbers==8.13.30
django-phonenumber-field==7.3.0

# SEO
django-meta==2.3.0
django-robots==6.1

# Utilities
Pillow==10.2.0
ics==0.7.2
pydantic==2.6.3

# Caching & Tasks (optional)
redis==5.0.3
celery==5.3.6

# Development & Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==24.3.0

# Code Quality
black==24.3.0
isort==5.13.2
ruff==0.3.3
mypy==1.9.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-redis==4.6.0.20240311

# Optional (commented out for MVP)
# sentry-sdk==1.43.0
# stripe==8.9.0
# boto3==1.34.69
# drf-spectacular==0.27.1


